% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNA.R
\name{generateNA}
\alias{generateNA}
\title{Generate Missing Values Completely at Random}
\usage{
generateNA(dat, pmiss = 0.2, seed = 123)
}
\arguments{
\item{dat}{complete data.frame.}

\item{pmiss}{proportion of NA.}

\item{seed}{seed value to ensure reproducibility.}
}
\value{
data.frame with the desired proportion of missing values.
}
\description{
Introducing a proportion of missing values given a dataframe.
}
\examples{

data(iris)
# introduce 30\% of NA
irisNA <- generateNA(iris, 0.3)
# check the proportion of NA 
 mean(is.na(irisNA))
}
